Added some ordering numbers to the title
On my flight, I decided that I want to read these posts in the order
that I thought it was most useful.

Since I didn’t have the order number, I couldn’t figure this out.

Reddit provides a way of knowing which of the posts are better (up
votes).

I have since added the order of the post

Added a config file/parser for ease of use